+++ src/main/java/com/github/blutorange/maven/plugin/closurecompiler/plugin/MojoMetadata.java
@@ -2,6 +2,7 @@ package com.github.blutorange.maven.plugin.closurecompiler.plugin;
 
 import org.apache.maven.plugin.logging.Log;
 import org.apache.maven.project.MavenProject;
+import org.sonatype.plexus.build.incremental.BuildContext;
 
 public interface MojoMetadata {
   MavenProject getProject();
@@ -10,5 +11,5 @@ public interface MojoMetadata {
 
   String getEncoding();
 
-  boolean isVerbose();
+  BuildContext getBuildContext();
 }